Collect coverage on all Pythons
Seems that Python 2 was getting left out of the mix. However sometimes
code needs to be special cased for Python 2. As these branches were not
being counted if tested on Python 2, it caused the coverage to dip
superficially. Hence this changes tox to track coverage on all Python
versions and report the results.