Switch from `buffer_copy` to `ndarray_copy`
The former has been replaced by the latter as of Numcodecs 0.6.0. That
said, this is technically internal spec. So maybe we shouldn't rely on
it here if we can avoid it. For now this just fixes the import/usage,
but we may want to replace this with a different strategy long term.