Ensure `DictStore` uses `bytes` to store blobs
The `DictStore` is pretty reliant on the fact that values are immutable
and can be easily compared. For example `__eq__` assumes that all
contents can be compared easily. This works fine if the data is `bytes`.
However it doesn't really work for `ndarray`s for example. Previously we
would have stored whatever the user gave us here. This means comparisons
could falldown in those cases as well (much as the example in the
tutorial has highlighted on CI). Now we effectively require that the
data be something that can either be coerced to `bytes` (e.g. via the
new/old buffer protocol) or is `bytes` to begin with. Make sure not to
force this requirement when nesting one `MutableMapping` within another.