Drop `test_getsize_ext`
This test case seems to be ill-posed. Anytime we store `object`s to
`Array`s we require an `object_codec` to be specified. Otherwise we have
no clean way to serialize the data. However this `DictStore` test breaks
that assumption by explicitly storing an `object` type in it even though
this would never work for the other stores (particularly when working
with `Array`s). This includes in-memory Zarr `Array`s, which would be
backed by `DictStore`. Given this, we go ahead and drop this test case.