avoid race condition during chunk write (#327)
* avoid race condition during chunk write

When the chunk file is first removed before the new version
is moved into place, racing reads may encounter a missing chunk.

Using rename() or replace() without remove() avoids the issue
on Posix-Systems as the methods are atomic. The fallback of
remove() -> rename() is included for Windows pre Python 3.3.

Fixes #263

* move feature-detection to init-time

so it's not repeated on every write

* use pyosreplace to get atomic replace() on Windows

* disable coverage count for legacy branches

* Use conditional instead of env marker

Because the env markers didn't work. Just guessing at this point.

* add pyosreplace package to requirements

* release notes [ci skip]