Cast datetime and timedelta to signed 64-bit int (#344)
* Cast datetime and timedelta to signed 64-bit int

The `NaT` type is represented as `-0`. As a result, casting to an
unsigned integral fails and throws an error. However casting to a signed
integral type does not have this problem and proceeds without issues.

* Update datetime/timedelta test to use signed ints

* Test encode/decode of a datetime/timedelta array

Use a structured array with datetime and timedelta values and a fill
value of NaT to test a bunch of different workarounds for encoding and
decoding datetime and timedelta values and array.

* Note improved `NaT` handling in the changelog

* Use `dtype` to cast `NaT` explicitly

* test NaT as fill_value for zarr.full()

* Drop extra whitespace

Should fix the flake8 error seen on CI.

ref: https://travis-ci.org/zarr-developers/zarr/jobs/465075634