Clear key-value pair if chunk is just fill value
Add a simple check to see if the key-value pair is just being set with a
chunk equal to the fill value. If so, simply delete the key-value pair
instead of storing a chunk that only contains the fill value. The Array
will behave the same externally. However this will cutdown on the space
require to store the Array. Also will make sure that copying one Array
to another Array won't dramatically effect the storage size.