Optionally use LZMA if it is available
LZMA is Python 3 only. That said, we do use a backport package for
Python 2 support. However it is a little tricky to install (particularly
on Windows and/or with pip), so we cannot always guarantee that we have
LZMA let alone that we can test it. While it would be nice to figure out
how to ensure LZMA is available in all of our testing environments, for
now simply try to test LZMA when it is available and ignore it
otherwise.