Ensure `str` before using `json.loads`
In Python 3.6+, `json.loads` will ensure data is handled with the proper
encoding even if it is in a binary format like a `bytes` object. Python
2.7's `str` is `bytes` so this already works there. However Python 3
versions pre-3.6 need a properly decoded `str`. Thus we fix the test to
use `ensure_str` to provide a `str` for use with `json.loads` to fix
this issue.