Create an SQLite-backed mutable mapping
Implements a key-value store using SQLite. As this is a builtin module
in Python and a common database to use in various languages, this should
have high utility and be very portable. Not to mention many databases
provide an SQLite language on top regardless of the internal
representation. So this can be a great template for users wishing to
work with Zarr in their preferred database.