Ensure all SQL commands are capitalized
Appears some of these commands work without capitalization. However as
the docs show commands as capitalized, ensure that we are doing the same
thing as well. That way this won't run into issues with different SQL
implementations or older versions of SQLite that are less forgiving.
Plus this should match closer to what users familiar with SQL expect.