Drop no longer needed flake8 error suppression
This was needed when the `import` of `sqlite3` was only here to ensure
that it existed (even though it wasn't used). Now we make use of
`sqlite3` where it is being imported. So there is no need to tell flake8
to not worry about the unused import as there isn't one.