Force some parameters to defaults
As some of these parameters no longer make sense to be user
customizable, go ahead and just set their values as part of the
`sqlite3.connect` call. This ensures that they are set the way we
expect. Also it ensures that if users try to mess with them, an error
will be raised due to duplicate keyword arguments.

To elaborate on why these parameters are not user configurable any more,
`detect_types` only makes sense if one is building their own table with
specific types. Instead we build the table for users and have very
generic types (i.e. text and blobs), which are not worth checking. As we
commit after every modification to the database to make it more friendly
for other threads and processes, the `isolation_level` might as well be
to auto-commit. Setting it to anything else really has no effect.
Finally there is no need for `check_same_thread` to be anything other
than `False` as we are guaranteeing everything is committed after
mutation, which ensures the database is thread and process safe.