Implement optimized `rmdir` for `SQLiteStore`
Provides an SQL implementation of `rmdir` that is a bit better optimized
for removing anything that matches the specified path as opposed to
doing multiple removals. If it is detected that the root directory is
being removed, simply fallback to clear, which is optimized for that use
case as it uses `DROP TABLE` instead of deleting rows. Otherwise remove
any path that begins with the normalized user-provided path as long as
it may contain at least one more character after. This stops `rmdir`
from removing a key-value pair where the key exactly matches normalized
user-provided path (i.e. not a "directory" as it contains data).