Implement optimized `getsize` for `SQLiteStore`
Take advantage of SQLite's ability to query and filter tables quickly to
implement `getsize` entirely in SQL (with the exception of path
normalization to sanitize user input). Measures the `LENGTH` of all
blobs in the column and calls `SUM` to get their aggregate size. In the
event that there are no matches, use `COALESCE` to replace the `NULL`
value returned by `SUM` with `0` instead.