Implement optimized `listdir` for `SQLiteStore`
Take advantage of SQLite's ability to query and filter tables quickly to
implement `listdir` entirely in SQL (with the exception of path
normalization to sanitize user input). Makes use of a nested
`SELECT`/`AS` to build a set of partial keys below top-level key. These
are then further split to get only the portion directly under the
top-level key and not any of their children.