Implement `rename` for `SQLiteStore`
Creates an SQL implementation of `rename` for `SQLiteStore`. As there
isn't a way to safely change the keys in `SQLiteStore` (since they are
`PRIMARY`), simply create a temporary table that copies over the
key-value pairs with keys renamed using a nested `SELECT` statement.
Then delete all key-value pairs that match the keys to move. Finally
copy all key-value pairs from the temporary table into our table and
delete the temporary table. Perform all of this as a transaction so only
the final result of the rename is visible to others.