Fuse `SELECT`s in `SQLiteStore`'s `__contains__`
Simplifies the SQL used in `SQLiteStore`'s `__contains__` method by
fusing the two `SELECT` statements into one. Does this by using
`COUNT(*)` to determine how many rows are left after the `SELECT`. As
the selection checks for an exact match with the key (and keys are
`PRIMARY`), there either is exactly `1` or `0`. So this works the same
as `SELECT EXISTS`, but with a single `SELECT` instead.