Just name the SQL table "zarr"
Instead of allowing the user to customize where the table is stored,
just set it to "zarr". This avoids issues with the table name
potentially exploiting injection attacks. Besides its unclear this level
of flexibility is really needed given Zarr supports Groups and thus can
store many Arrays in the same key-value store.