Check SQLite Cursor's rowcount for deletion
Instead of checking if a particular key exists and then either raising a
`KeyError` or deleting it in `SQLiteStore`, go ahead with the deletion
and check the value of `rowcount`. As the keys are primary, they must be
unique and thus each one only occurs once. Thus if deletion worked, the
`rowcount` will be exactly `1` (it cannot be larger). Alternatively if
deletion failed, the `rowcount` would be `0`. Thus we can simply check
if the `rowcount` is not `1` and raise the `KeyError`. This should
improve the performance a bit.