Test the `popitem` method of `MutableMapping`s too
Adds a test for the `popitem` method of `MutableMapping`'s as well.
Typically this is not used, but it does get used in the default `clear`
method. Given this method doesn't guarantee an order, use a store with a
single-key value pair to simplify the test logic. Also test the case of
an empty store to make sure it errors out properly. Override the
`popitem` method appropriate for stores that do not handle removal
properly.