Add a test to check that all values are bytes
Currently this is a somewhat loose requirement that not all stores
enforce. However it is a useful check to have in some cases.
Particularly this is a useful constraint with in-memory stores where
there is a concern that the data in the store might be manipulated
externally due to ndarray views. The bytes instances don't have this
problem as they are immutable and own their data. While views can be
take onto bytes instances, they will be read-only views and thus are not
a concern.

For other stores that place their data in some other storage backend
(e.g. on disk), this is less of a concern. Also other stores may choose
to represent their data in other ways (e.g. LMDB with `memoryview`s).
Manipulating the loaded data from these stores is less of a concern
since as it doesn't affect their actually contents, only an in-memory
representation. In these cases, it may make sense to disable this test.