Disable bytes value test for LMDBStore w/buffers
It's possible to request that `LMDBStore` return `buffer`s/`memoryview`s
instead of returning `bytes` for values. This is incredibly useful as
LMDB is memory-mapped. So this avoids a copy when accessing the data.
However this means it will fail `test_store_has_bytes_values`. Though
this is ok as noted previously since that is not a hard requirement of
stores. So we just disable that test for this `LMDBStore` case. The
other `LMDBStore` case returns `bytes` instances instead (copying the
data); so, it passes this test without issues.