Have CustomMapping ensure it stores bytes
Since people will coming looking to see how a store should be
implemented, we should show them good behavior. Namely we should ensure
the values provided to `__setitem__` are `bytes`-like and immutable
(once stored). By coercing the data to `bytes` we ensure that the data
is `bytes`-like and we ensure the data is immutable since `bytes` are
immutable.