Disable bytes value test for LRUStoreCache
As the point of the `LRUStoreCache` is to merely hold onto values
retrieved from the underlying store and keep them around in memory
unaltered, the caching layer doesn't have any control over what type of
values are returned to it. Thus it doesn't make much sense to test
whether the values it returns are of `bytes` type or not. Though this is
fine as there is not strict requirement that values of `bytes` type be
returned by stores, so simply disable `test_store_has_bytes_values` for
the `LRUStoreCache` test case with a note explaining this.