Ensure `bytes` in `_encode_chunk` for `dict`
In `Array` when running `_encode_chunk` and using a `dict`-based chunk
store, ensure that the chunk data is of `bytes` type. This is done to
convert the underlying data to an immutable value to protect against
external views onto the data from changing it (as is the case with NumPy
arrays). Also this is done to ensure it is possible to compare
`dict`-based stores easily.

While coercing to a `bytes` object can introduce a copying, this
generally won't happen if a compressor is involved as it will usually
have returned a `bytes` object. Though filters may not, which could
cause this to introduce an extra copy if only filters and no compressors
are used. However this is an unlikely case and is not as important as
guaranteeing the values own their own data and are read-only. Plus this
should allow us to drop the preventive copy that happens earlier when
storing values as this neatly handles that case of no filters and no
compressors.