Only catch `OSError` if its `errno` equal `EEXIST`
On Python 3, we would only want to catch the race case here by checking
for the `FileExistsError` and suppressing raising for that case. However
there is no `FileExistsError` on Python 2. Only `OSError` exists on
Python 2/3. However that catches some other errors that we might not
want to suppress here. So we check the `errno` of the `OSError` instance
to verify it is caused by an existing file/directory represented by the
`EEXIST` error number. This amounts to catching the `FileExistsError` in
a Python 2/3 friendly way. All other `OSErrors` we raise as before.