Use warnings instead of logging
As we are only using the logger to issue warnings, just issue warnings
instead and drop the logger. Picked `RuntimeWarning`s for these two
warnings as they are in regard to "dubious runtime behavior". Namely
using compressors that are not explicitly supported by many N5
implementations to write N5 files.