Test that N5Store raises for certain attributes
As N5 stores attributes and metadata in the same JSON object, there are
some keys that are simply not allowed to be attributes as they would
conflict with the metadata (and possibly corrupt the data if they were
written in). The `N5Store` correctly raises for these cases, but this
was not being tested previously. Here we add a test to cover this case.