Optionally memory-mapping data in `DirectoryStore`
This provides users the option to memory-map the data from the
`DirectoryStore` instead of reading it all into memory at once. This can
be useful if the chunks are especially large. Also can be useful in
combination with zero-copy changes made to Numcodecs to decompress data
straight from disk into a decompressed buffer. In cases where the data
is not compressed, this can make accessing pieces of chunks pretty
efficient.

To make resource management simple, we leverage NumPy's `memmap`, which
cleans up after itself once no more references are held (e.g. RAII). So
there is no need to explicitly close the file. This way if the
memory-mapped file is only used for a single operation and tossed, it
will automatically be cleaned up without needing other code outside the
store to handle this.

So as to ensure no inadvertent changes are made to the store, the
memory-mapped file is opened in read-only mode. Thus we disallow
accidental overwrites of data on disk. This constraint applies to any
views taken of the `memmap` (e.g. NumPy `ndarray`s or `memoryview`s). Of
course users can still use `__setitem__` to overwrite the data on disk.

At present, memory-mapping only occurs when reading data from disk. It
is not used to write data to disk. With reading data from disk, there
are clear benefits (as noted above). However it is unclear what the
benefits are for writing data to disk using memory-mapping. Particularly
as we completely overwrite data on disk in an atomic fashion, which
would be difficult to reproduce using a memory-mapped file.

Other stores inheriting from the `DirectoryStore` can benefit from this
change if they also pass along a `memmap` keyword argument to the
constructor and reuse its `__getitem__` either as is or in their own
subclass implementation of the method. No further changes are required.

Currently memory-mapping is optional and is opt-in. If it is seen to be
generally useful, we may make it opt-out instead. Though the option
remains useful should users find this does or does not make sense in
different contexts. Users could hack the value of `memmap` should they
need to on a case-by-case basis (e.g. only used for specific values),
but it is not necessarily recommended.