Implement `pop` and `popitem` for `DirectoryStore`
As memory-mapped files will become invalid once deleted,
`DirectoryStore`'s `pop` and `popitem` need to be implemented such that
they don't hold invalid references to files removed from the disk. To
handle this, simply call `ensure_bytes` on any value retrieved to ensure
it is copied into memory. This will be a no-op if memory-mapping was
disabled as the value would already be a `bytes` instance.