Add `ipytree` as an extra requirement
For users wishing to have a nice representation of groups in Jupyter
notebooks, `ipytree` is recommended to make these look a bit nicer as
well as enable interactivity. Though it is not a requirement as we can
always fallback to line representations. We make `ipytree` an optional
dependency as there is a little more work to install it depending on the
use case. So we want users to handle this intentionally.