Use `ensure_ndarray` on `out`
Since we are interested in getting an `ndarray` representing the buffer
within `out` for writing into, go ahead and use `ensure_ndarray` to
coerce the underlying buffer into an `ndarray`. This way we can avoid a
needless check and just write into any array-like value for `out` that
is provided.