Handle `out` when it doesn't supply a buffer
Appears that `out` can also be a Zarr `Array` or any other array-like
that does not expose a buffer, but does allow writing into. In these
cases `ensure_ndarray` will fail as there is not an underlying buffer
that can be used with an `ndarray`. To also handle this case, catch the
`TypeError` that `ensure_ndarray` will raise in this case and use that
to indicate whether `out` is now an `ndarray` or not. This allows us to
continue to write into arbitrary buffers, but also correctly handle
objects that do not expose buffers.