Refactor out `_tofile`/`_fromfile` from `DirectoryStore` (#503)
* Add `_fromfile` and `_tofile` methods

To make it easier to sub out different ways of reading and writing data,
create to methods to handle reading and writing directly. This way users
don't need to reimplement all of the same logic that `__getitem__` and
`__setitem__` are doing and can focus on simply reading and writing. Can
be useful for trying memmaping for example.

* Use _fromfile/_tofile in __getitem__/__set item__

* Update zarr/storage.py

Co-Authored-By: James Bourbeau <jrbourbeau@users.noreply.github.com>

* Add docstrings to `_fromfile` and `_tofile`

* No need to return in `_tofile`

* Adds release notes entry