Include `_meta_array` when pickling `Array`
To make sure that libraries that rely on pickling objects (like Dask or
other parallel libraries) respect the `_meta_array` attribute, pass it
along as part of the object's state when `__getstate__` is called. Since
`__setstate__` just calls the constructor with all arguments, we can be
sure that `_meta_array` will then be passed back into the new `Array`.
Though this all assumes that the `_meta_array` is something that can be
pickled.