Update Deprecating Warning with version number.
Adding the versing number make it slightly easier on consumers to update
to the new naming as they will get a sens of compatibility.

Also update pytest to fail is any deprecation warning are not marked in
the test suite, and catch them in more places.

In particular we realise that many more places trigger the warnings in
create_store so we catch the warning there, but also that
`pickle.load()` can trigger it.