Move metadata handling in a class.
This will make it a tiny bit easier to make a transition to V3 by having
groups/array use an instance to encode/decode instead of refering to the
function/module.

While for v2 it does not make much difference; I'm expecting instances
in V3 to likely be (some of) the pieces that handle extensions, and be
passed around in a hierarchy.

I currently re-expose the methods at the module level to not change the
API/tests, though I expect to wrap them and add deprecation once I've
move the Groups/Arrays/Attributes to use those.