Reduce number of store queries in open(mode='a').
In case where 'shape' is passed as argument, this will avoid calling
"contains_array" which does store queries, as well as avoid testign if
the array contains a group as anyway this is what we were going to do in
the else case.

This does change behavior in edge case:
 - if shape is passed and the key is a group, we'll try to open as array
 instead of groups.
 - if contains_array or contains_group would have raise we might skip
 over those checks.

Though both of the above case should not be found in user code and
are bugs AFAICT, and would just fail differently.