Remove unused default+ keyword that is only used internally
Move it to the only two places where it is used;
technically this is a change of API, but the function
`normalize_store_arg` is only used internally.

This make it a bit easier to follow what might happens when it is
called; in particular this will let us be sure that it always returns a
Store as as long as it is passed is a Store or a string, whicl before it
could have returned anything depending on the value of default.