Fix DirectoryStore (#773)
* Drop skip_if_nested_chunks from test_storage.py

* Add failing nested test

* Make DirectoryStore dimension_separator aware

* Migrate key logic to core rather than storage

Previous tests (now commented out) used logic in the store
classes to convert "0/0" keys into "0.0" keys, forcing the
store to be aware of array details. This tries to swap the
logic so that stores are responsible for passing dimension
separator values down to the arrays only. Since arrays can
also get the dimension_separator value from a .zarray file
they are now in charge.

* Fix linting in new test

* Extend the test suite for dim_sep

* Try fsspec 2021.7 (see #802)

* Revert "Try fsspec 2021.7 (see #802)"

This reverts commit 68adca50b62441dabc6b3f48364fe3dcf35eeb69.

* Fix N5Store

* Re-activate contested N5 test