Pass dimension_separator on fixture generation (fix #858)
Under some test conditions (conda-forge, Debian builds), the
fixtures directory is not available and is then re-created.
When dimension_separator is not passed to DirectoryStore, then
no metadata is assigned to the file. For the "flat" (as opposed
to "flat_legacy") fixture, this meant that the NestedDirectoryStore
did not correct its behavior leading to a failure.

see:
 - https://github.com/conda-forge/zarr-feedstock/pull/56
 - https://github.com/zarr-developers/zarr-python/issues/858