Add LGTM.com / DeepSource.io configuration files (#909)
* Add LGTM.com configuration file

The YAML configuration file for the LGTM static analysis tool can be
either lgtm.yml or .lgtm.yml:
  https://help.semmle.com/lgtm-enterprise/user/help/lgtm.yml-configuration-file.html

There is no need to integrate the LGTM tool in CI, as LGTM appears to
be running on all repositories it has been run on once. The results
currently appear here:
  https://lgtm.com/projects/g/zarr-developers/zarr-python

* Add DeepSource.io configuration file

The TOML configuration file is .deepsource.toml:
  https://deepsource.io/docs/concepts/#deepsourcetoml-file

DeepSource.io analysis must be enabled by a repository owner.

It doesn't look like it is possible to run DeepSource.io on each PR,
rather the monitored branch is analysed periodically, after commits
are pushed.