Fix a TOCTOU race in `open_array(mode="a")`
If you have multiple writers trying todo `zarr.open_array(..., "a")`,
the old code code sometimes raise a `ContainsArrayError`, even thogh
opening an already-existing array should be fine.

The underlying problem is a time-of-check/time-of-use race condition,
where the array can be created (by another writer) between the
`contains_array` check and the `init_array` call.

To fix this, we simply swallow the error from the `init_array` call.