Merge pre-release sections
An odd bug (or user error?!) with readthedocs leads
to all pre-release anchors in the changelog to be
reduced to just the final component (e.g. "a2").
Rather than struggle with this, this proposes to use
the official anchor for the upcoming release, but to
clearly mark it as a pre-release.