Update and document GPU buffer handling (#2751)
* Update GPU handling

This updates how we handle GPU buffers. See the new docs page for a
simple example.

The basic idea, as discussed in ..., is to use host buffers for all
metadata objects and device buffers for data.

Zarr has two types of buffers: plain buffers (used for a stream of
bytes) and ndbuffers (used for bytes that represent ndarrays). To make
it easier for users, I've added a new config option
`zarr.config.enable_gpu()` that can be used to update those both. If
we need additional customizations in the future, we can add them here.

* fixed doc

* Fixup

* changelog

* doctest, skip

* removed not gpu

* assert that the type matches

* Added changelog notes

---------

Co-authored-by: Davis Bennett <davis.v.bennett@gmail.com>