R1717: Consider using a dictionary comprehension
Emitted when we detect the creation of a dictionary using the dict()
callable and a transient list. Although there is nothing syntactically
wrong with this code, it is hard to read and can be simplified to a dict
comprehension. Also it is faster since you don't need to create another
transient list.