Provide pytest-plugin that registers a 'gpu' marker. (#3258)
https://github.com/zarr-developers/zarr-python/pull/3191/files#diff-a801b931f5dd8cd7df666e3f9936ff520669296303d3a3e469514443211885bdR43
changed `zarr.testing.utils` to use a top-level `pytest.mark.gpu`. Many
projects, including zarr, run with `--strict-markers` which causes
pytest to raise an error if the test you're running uses an unregistered
marker. This didn't cause any issues for zarr, since we already
registered the `gpu` marker.

However, downstream projects are intended to run the `StoreTests`, which
exposes them to this marker. To ensure they get it when installing zarr,
we'll include an entrypoint with pytest, so that pytest calls our code
to register this marker when these tests run.