Add exception for swift in fsspec path handling
http(s) and swift filesystems need the scheme in the path.
This commit adds a special case for the swift filesystem,
when checking for the scheme presence.

Also remove one if for the http(s) schemes. Here all schemes can
call fs._strip_protocol(path). For the filesystems that
need the scheme in the path, this function will not remove it.