optimize shard writing
Writing to sharded arrays was up to 10x slower for largish chunk sizes
because the _ShardBuilder object has many calls to np.concatenate. This
commit coalesces these into a single concatenate call, and improves write
performance by a factor of 10 on the benchmarking script in #3560.

Added a new core.Buffer.combine API

Resolves #3560

Signed-off-by: Noah D. Brenowitz <nbren12@gmail.com>