Refactor Codec interface
This refactors the interface of `Codec.read` and `Codec.write` to
move from an `Iterable[tuple[...]` to an `Iterable[BatchInfo]`.

Two things motivate this change

1. Readability: I struggle to remember what the 4th member of these
   complex tuples. Having the name `info.out_selection` to remind
   me is helpful.
2. Possible future-proofing: right now, any change to the interface
   is a hard break since the number of elements in the tuple will
   change. There may be a class of changes to the interface where
   we can add additional information to `BatchInfo` without breaking
   backwards compatibility.

I don't want to oversell motivaiton 2 though. If something is important
enough to add to the interface, then presumably we expectd implementations
to, you know, use it.